requirements.txt
absl-py==1.1.0
avalanche-lib==0.1.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
docker-pycreds==0.4.0
entrypoints==0.4
filelock==3.7.1
fonttools==4.29.1
gdown==4.4.0
gensim==4.1.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.6
google-auth-oauthlib==0.4.6
GPUtil==1.4.0
grpcio==1.46.3
idna==3.3
imbalanced-learn==0.9.0
importlib-metadata==4.11.1
ipykernel==6.13.1
ipython==7.33.0
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.0.6
jupyter-core==4.10.0
kiwisolver==1.3.2
lxml==4.8.0
Markdown==3.3.7
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
nltk==3.7
numpy==1.21.5
oauthlib==3.2.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==21.2.2
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
Pygments==2.12.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-adapt==0.0.74
pytorch-metric-learning==1.4.0
pytorchcv==0.0.67
pytz==2022.1
PyYAML==6.0
pyzmq==19.0.2
quadprog==0.1.11
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
sentry-sdk==1.5.6
setuptools==58.0.4
shortuuid==1.0.8
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
soupsieve==2.3.2.post1
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl==3.1.0
timm==0.6.7
torch==1.8.1+cu102
torch-tb-profiler==0.4.0
torchaudio==0.8.1
torchmetrics==0.9.0
torchvision==0.9.1+cu102
tornado==6.1
tqdm==4.62.3
traitlets==5.2.2.post1
typing_extensions==4.1.1
urllib3==1.26.8
wandb==0.12.10
wcwidth==0.2.5
Werkzeug==2.1.2
wheel==0.37.1
yaspin==2.1.0
zipp==3.7.0
